projects/capstone/starter/requirements.txt
alembic==1.7.1
aniso8601==6.0.0
astroid==2.2.5
attrs==21.2.0
Babel==2.9.0
cffi==1.14.6
click==8.0.1
cryptography==3.4.7
ecdsa==0.13.2
Flask==2.0.1
Flask-Cors==3.0.8
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.17.1
greenlet==1.1.1
gunicorn==20.0.4
importlib-resources==5.2.2
iniconfig==1.1.1
isort==4.3.18
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.4.0
linecache2==1.0.0
Mako==1.1.5
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
pluggy==0.13.1
psycopg2-binary==2.8.2
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.3.1
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.6.0
python-jose==3.3.0
python-jose-cryptodome==1.3.2
pytz==2019.1
rsa==4.7.2
six==1.12.0
SQLAlchemy==1.4.23
toml==0.10.2
traceback2==1.4.0
typed-ast==1.4.2
unittest2==1.1.0
Werkzeug==2.0.1
wrapt==1.11.1
WTForms==2.3.3
zipp==3.5.0
